Revert "Merge tag 'LA.UM.7.3.r1-07900-sdm845.0' of https://source.codeaurora.org/quic/la/platform/vendor/qcom-opensource/wlan/qcacld-3.0 into HEAD"

+++ drivers/staging/qcacld-3.0/core/sme/inc/csr_api.h
@@ -1594,7 +1594,7 @@ typedef struct tagCsrRoamInfo {
 	int rx_rate;
 	tSirMacCapabilityInfo capability_info;
 	uint32_t rx_mc_bc_cnt;
-	uint16_t roam_reason;
+	uint8_t roam_reason;
 } tCsrRoamInfo;
 
 typedef struct tagCsrFreqScanInfo {
